feat(copyright): Manage translations in config
As copyright is part of config, it is easier to manage its translations
at the same place.
`config.extra.translate_copyright` is still used to declare translation
intent for backward compatibility.
It allows to raise an error at build time if someone forgot to move from
using i18n files as previously required to the config.toml.